🚚build : 구조 변경
