+++ apps/user/utils/format.py
@@ -1,0 +1,2 @@
+def format_phone(phone):
+    return f"{phone[:3]}-{phone[3:7]}-{phone[7:]}"
